refactor: consolidate examples from 27 to 23
Merge overlapping examples to reduce cognitive overhead for new users:

- basic.rs + stdio_server.rs -> getting_started.rs
  Covers tools, resources, prompts, and stdio transport in one file

- transport_middleware.rs + tool_middleware.rs + middleware_showcase.rs -> middleware.rs
  Shows all four middleware patterns: transport-level, per-tool, per-resource,
  per-prompt, and guards -- in one cohesive example

- http_client_cli.rs -> http_client.rs (renamed for discoverability)

- dynamic_tools.rs + skill_prompts.rs -> dynamic_capabilities.rs
  Covers DynamicToolRegistry, DynamicPromptRegistry, and
  DynamicResourceRegistry in one example

The middleware.rs example is the first to demonstrate per-resource and
per-prompt .layer() middleware, filling a coverage gap.

Cargo.toml example entries reorganized into logical sections:
getting started, transports, middleware, authentication, clients,
sampling, dynamic capabilities, advanced patterns, macros.

Refs #754
